+++ src/mon-make.h
@@ -37,6 +37,8 @@ struct monster *place_pet_beside_player(struct monster_race *race);
 int16_t place_monster(struct chunk *c, struct loc grid, struct monster *mon,
 	uint8_t origin);
 int mon_hp(const struct monster_race *race, aspect hp_aspect);
+bool monster_race_fits_grid(struct chunk *c, struct loc grid,
+	const struct monster_race *race);
 bool place_new_monster(struct chunk *c, struct loc grid,
 	struct monster_race *race, bool sleep, bool group_ok,
 	struct monster_group_info group_info, uint8_t origin);